requirements.txt
# ============================================
# Brain RH Migration - Requirements complets
# ============================================
# Projet: Système de matching CV-Offre avec IA
# Stack: FastAPI (backend) + React (frontend)
# Date: 27 octobre 2025
# ============================================

# ===================
# API Web Framework
# ===================
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
sqlmodel==0.0.22
python-multipart==0.0.6  # Upload de fichiers (CV PDF/DOCX)

# ===================
# IA et LLM
# ===================
openai==1.63.2                  # OpenAI API (GPT-5 mini)
sentence-transformers==5.1.1    # Embeddings sémantiques
numpy>=1.24.0                   # Calculs vectoriels
scikit-learn>=1.3.0             # Outils ML (cosine similarity)

# ===================
# Parsing et validation
# ===================
python-dateutil>=2.8.0          # Parsing dates (experience_analyzer)
jsonschema==4.23.0              # Validation JSON schemas
pyyaml>=6.0                     # Configuration YAML

# ===================
# Extraction documents
# ===================
PyMuPDF>=1.23.0                 # Extraction PDF
docx2txt>=0.8                   # Extraction DOCX

# ===================
# Export et HTTP
# ===================
openpyxl==3.1.5                 # Export Excel
requests>=2.31.0                # HTTP client (xAI API)
httpx>=0.24.0                   # HTTP async (tests FastAPI)

# ===================
# Utilitaires
# ===================
python-dotenv==1.0.0            # Variables d'environnement
tenacity>=8.2.0                 # Retry logic (robustesse API)

# ===================
# Tests (optionnel mais recommandé)
# ===================
pytest>=7.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0                 # Coverage des tests

# ===================
# Notes de déploiement
# ===================
# Frontend (React): voir frontend/package.json
# - Vite, React 19, TypeScript, Tailwind CSS
# - À builder avec: cd frontend && npm run build
#
# Base de données: SQLite (intégré)
# - Pas de dépendance externe supplémentaire
#
# Variables d'environnement requises (.env):
# - OPENAI_API_KEY: Clé API OpenAI
# - XAI_API_KEY: Clé API xAI (Grok, pour fiches d'entretien)
